+++ src/main/java/org/ezcode/codetest/domain/community/repository/DiscussionRepository.java
@@ -1,11 +1,11 @@
 package org.ezcode.codetest.domain.community.repository;
 
+import java.util.List;
 import java.util.Optional;
 
 import org.ezcode.codetest.domain.community.dto.DiscussionQueryResult;
 import org.ezcode.codetest.domain.community.model.entity.Discussion;
 import org.ezcode.codetest.domain.language.model.entity.Language;
-import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
 
 public interface DiscussionRepository {
@@ -14,10 +14,14 @@ public interface DiscussionRepository {
 
 	Optional<Discussion> findById(Long discussionId);
 
-	Page<DiscussionQueryResult> findAllByProblemId(Long problemId, String sortBy, Long userId, Pageable pageable, Long ttt);
+	List<Long> findDiscussionIdsByProblemId(Long problemId, String sortBy, Pageable pageable);
+
+	List<DiscussionQueryResult> findDiscussionsByIds(List<Long> discussionIds, Long currentUserId);
 
 	void updateDiscussion(Discussion discussion, Language language, String content);
 
 	void deleteDiscussion(Discussion discussion);
+
+	Long countByProblemId(Long problemId);
 
 }
