+++ src/main/java/org/ezcode/codetest/domain/community/service/DiscussionDomainService.java
@@ -1,5 +1,10 @@
 package org.ezcode.codetest.domain.community.service;
 
+import java.util.List;
+import java.util.Map;
+import java.util.function.Function;
+import java.util.stream.Collectors;
+
 import org.ezcode.codetest.domain.community.dto.DiscussionQueryResult;
 import org.ezcode.codetest.domain.community.exception.CommunityException;
 import org.ezcode.codetest.domain.community.exception.CommunityExceptionCode;
@@ -7,6 +12,7 @@ import org.ezcode.codetest.domain.community.model.entity.Discussion;
 import org.ezcode.codetest.domain.community.repository.DiscussionRepository;
 import org.ezcode.codetest.domain.language.model.entity.Language;
 import org.springframework.data.domain.Page;
+import org.springframework.data.domain.PageImpl;
 import org.springframework.data.domain.Pageable;
 import org.springframework.stereotype.Service;
 
@@ -37,9 +43,23 @@ public class DiscussionDomainService {
 		return discussion;
 	}
 
-	public Page<DiscussionQueryResult> getAllDiscussionsByProblemId(Long problemId, String sortBy, Long userId, Pageable pageable, Long ttt) {
+	public Page<DiscussionQueryResult> getAllDiscussionsByProblemId(Long problemId, String sortBy, Long userId, Pageable pageable) {
 
-		return discussionRepository.findAllByProblemId(problemId, sortBy, userId, pageable, ttt);
+		// 성능 향상 위해 쿼리 분리
+		List<Long> discussionIds = discussionRepository.findDiscussionIdsByProblemId(problemId, sortBy, pageable);
+		List<DiscussionQueryResult> results = discussionRepository.findDiscussionsByIds(discussionIds, userId);
+
+		Long totalCount = discussionRepository.countByProblemId(problemId);
+
+		// `WHERE IN` 절은 ID 목록의 순서를 보장하지 않으므로, 처음에 정렬해서 얻은 ID 목록의 순서대로 results를 다시 정렬
+		Map<Long, DiscussionQueryResult> resultMap = results.stream()
+			.collect(Collectors.toMap(DiscussionQueryResult::getDiscussionId, Function.identity()));
+
+		List<DiscussionQueryResult> sortedResults = discussionIds.stream()
+			.map(resultMap::get)
+			.collect(Collectors.toList());
+
+		return new PageImpl<>(sortedResults, pageable, totalCount);
 	}
 
 	public Discussion modify(Long discussionId, Long problemId, Long userId, Language language, String content) {
