+++ src/main/java/org/ezcode/codetest/infrastructure/persistence/repository/community/discussion/DiscussionJpaRepository.java
@@ -3,5 +3,5 @@ package org.ezcode.codetest.infrastructure.persistence.repository.community.disc
 import org.ezcode.codetest.domain.community.model.entity.Discussion;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface DiscussionJpaRepository extends JpaRepository<Discussion, Long>, DiscussionQueryRepository, DiscussionQueryOptimizedRepository {
+public interface DiscussionJpaRepository extends JpaRepository<Discussion, Long>, DiscussionQueryRepository {
 }
