+++ src/main/java/org/ezcode/codetest/infrastructure/persistence/repository/community/discussion/DiscussionQueryRepository.java
@@ -3,15 +3,14 @@ package org.ezcode.codetest.infrastructure.persistence.repository.community.disc
 import java.util.List;
 
 import org.ezcode.codetest.domain.community.dto.DiscussionQueryResult;
-import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
 
 public interface DiscussionQueryRepository {
 
-	Page<DiscussionQueryResult> findAllByProblemId(Long problemId, String sortBy, Long currentUserId, Pageable pageable);
+	List<Long> findDiscussionIdsByProblemId(Long problemId, String sortBy, Pageable pageable);
 
-	List<Long> findDiscussionIdsByProblemIdWithSubquery(Long problemId, String sortBy, Pageable pageable);
+	List<DiscussionQueryResult> findDiscussionsByIds(List<Long> discussionIds, Long currentUserId);
 
-	List<DiscussionQueryResult> findDiscussionsByIdsWithSubquery(List<Long> discussionIds, Long currentUserId);
+	Long countByProblemId(Long problemId);
 
 }
