+++ src/main/java/org/ezcode/codetest/presentation/community/DiscussionController.java
@@ -46,7 +46,6 @@ public class DiscussionController {
 	public ResponseEntity<Page<DiscussionResponse>> getDiscussions(
 		@PathVariable Long problemId,
 		@RequestParam(defaultValue = "best") String sortBy,
-		@RequestParam Long ttt,
 		@PageableDefault Pageable pageable,
 		@AuthenticationPrincipal AuthUser authUser
 	) {
@@ -55,7 +54,7 @@ public class DiscussionController {
 
 		return ResponseEntity
 			.ok()
-			.body(discussionService.getDiscussions(problemId, sortBy, currentUserId, pageable, ttt));
+			.body(discussionService.getDiscussions(problemId, sortBy, currentUserId, pageable));
 	}
 
 	@PutMapping("/{discussionId}")
