+++ src/test/resources/application-test.properties
@@ -1,7 +1,7 @@
 # ========================
 # Spring Config
 # ========================
-spring.application.name=min
+spring.application.name=ezcode
 spring.config.import=optional:file:.env[.properties]
 
 # ========================
